feat-gallery완 refactor2차 완료
